requirements.txt
streamlit>=1.35.0
anthropic>=0.25.0
feedparser>=6.0.11
python-dotenv>=1.0.0
requests>=2.31.0
streamlit-mic-recorder>=0.0.8
edge-tts>=6.1.9
ddgs>=0.1.0
